% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DR_exposure.R
\name{DR_exposure}
\alias{DR_exposure}
\title{Density Ranking exposure}
\usage{
DR_exposure(
  x,
  day = NULL,
  cellsize = 100,
  smoothing = 0.0007,
  normalize = TRUE,
  env_data = NULL,
  data_extent = NULL,
  stats = NULL,
  act_and_env = FALSE
)
}
\arguments{
\item{x}{data frame with lon lat coordinate columns}

\item{day}{string in date format compatible with date column in x}

\item{cellsize}{size of raster cell in meters}

\item{smoothing}{smoothing parameter}

\item{normalize}{argument if activity data should be normalized to 0-1 values range}

\item{env_data}{SparRaster object of envirinmental data}

\item{data_extent}{TODO}

\item{stats}{statistics calculated}

\item{act_and_env}{TODO}
}
\value{
list of SpatRaster result and list of statistics
}
\description{
Using DR function from Yen-Chi Chen's repository density_ranking, only in WGS84 with smoothing parameter
}
