% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure_KDE.R
\name{exposure_KDE}
\alias{exposure_KDE}
\title{Kernel Density Estimation exposure (TDA)}
\usage{
exposure_KDE(
  x,
  cellsize = NULL,
  bandwidth = 200,
  env_data = NULL,
  scale_01 = FALSE,
  data_extent = NULL,
  input_crs = "EPSG:4326",
  output_crs = NULL,
  stats = NULL
)
}
\arguments{
\item{x}{data frame with lon lat coordinates columns}

\item{cellsize}{positive numeric size of raster cells in meters}

\item{bandwidth}{positive numeric bandwidth in units of x}

\item{env_data}{SpatRaster object of environmental data}

\item{scale_01}{boolean argument if activity data should be rescaled to 0-1 values range}

\item{data_extent}{TODO}

\item{input_crs}{character or terra crs object specifying coordinate reference system of coordinates in x data frame}

\item{output_crs}{character or terra crs object of coordinate reference system of output}

\item{stats}{vector of characters statistics to be calculated. See terra::global. "count", "range" and "area" additionally added.}
}
\value{
list of SpatRaster result and data frame of statistics
}
\description{
Kernel Density Estimation method activity space and environmental exposure. Using kde() function from TDA package.
Bandwidth is a smoothing parameter same as h in TDA::kde(). In order to receive activity space ignore env_data argument.
}
\examples{

statistics = c("count", "area", "min", "max", "range", "mean", "std", 'sum')


# activity space
exposure_KDE(x = geolife_sandiego, cellsize = 50, bandwidth = 25,
 scale_01 = TRUE, input_crs = "EPSG:4326", output_crs = "EPSG:32611",
 stats = statistics)

#environmental exposure

ndvi_data = terra::rast(system.file("extdata/landsat_ndvi.tif", package = "twsagps"))

exposure_KDE(x = geolife_sandiego, cellsize = 50, bandwidth = 25,
 env_data = ndvi_data, scale_01 = TRUE, input_crs = "EPSG:4326",
 output_crs = "EPSG:32611", stats = statistics)



}
