% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure_LS.R
\name{exposure_LS}
\alias{exposure_LS}
\title{Line Segment Method exposure}
\usage{
exposure_LS(
  x,
  time_data = NULL,
  time_unit = "mins",
  cellsize = NULL,
  bandwidth = 200,
  env_data = NULL,
  data_extent = NULL,
  input_crs = "EPSG:4326",
  output_crs = NULL,
  stats = NULL
)
}
\arguments{
\item{x}{data frame with lon lat coordinates columns}

\item{time_data}{name of column in x containing POSIXct data class}

\item{time_unit}{character indicating unit of weights of time_data. Ignored if time_data not specified. See difftime()}

\item{cellsize}{positive numeric size of raster cells in meters}

\item{bandwidth}{positive numeric size of segments in meters}

\item{env_data}{SpatRaster object of environmental data}

\item{data_extent}{TODO}

\item{input_crs}{character or terra crs object specifying coordinate reference system of coordinates in x data frame}

\item{output_crs}{character or terra crs object of coordinate reference system of output}

\item{stats}{vector of characters statistics to be calculated. See terra::global. "count", "range" and "area" additionally added.}
}
\value{
list of SpatRaster result and data frame of statistics
}
\description{
Line Segment method activity space and environmental exposure. Based on code from Micheal garber's repository microclim-static-v-dynam.
In order to receive activity space ignore env_data argument.
}
\examples{

statistics = c("count", "area", "min", "max", "range", "mean", "std", 'sum')


# activity space
exposure_LS(x = geolife_sandiego, time_data = dateTime, time_unit = "mins",
 cellsize = 50, bandwidth = 200, input_crs = "EPSG:4326", output_crs = "EPSG:32611",
  stats = statistics)

#environmental exposure

ndvi_data = terra::rast(system.file("extdata/landsat_ndvi.tif", package = "twsagps"))

exposure_LS(x = geolife_sandiego, time_data = dateTime, time_unit = "mins",
 cellsize = 50, bandwidth = 200, env_data = ndvi_data, input_crs = "EPSG:4326",
 output_crs = "EPSG:32611", stats = statistics)


}
