% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure_PO.R
\name{exposure_PO}
\alias{exposure_PO}
\title{Point Overlay exposure}
\usage{
exposure_PO(
  x,
  cellsize = NULL,
  env_data = NULL,
  scale_01 = FALSE,
  data_extent = NULL,
  input_crs = "EPSG:4326",
  output_crs = NULL,
  stats = NULL
)
}
\arguments{
\item{x}{data frame with lon lat coordinates columns}

\item{cellsize}{positive numeric size of raster cells in meters}

\item{env_data}{SpatRaster object of environmental data}

\item{scale_01}{boolean argument if activity data should be rescaled to 0-1 values range}

\item{data_extent}{TODO}

\item{input_crs}{character or terra crs object specifying coordinate reference system of coordinates in x data frame}

\item{output_crs}{character or terra crs object of coordinate reference system of output}

\item{stats}{vector of characters statistics to be calculated. See terra::global. "count", "range" and "area" additionally added.}
}
\value{
list of SpatRaster result and data frame of statistics
}
\description{
Point Overlay method activity space and environmental exposure. In order to receive activity space ignore env_data argument.
}
\examples{

statistics = c("count", "area", "min", "max", "range", "mean", "std", 'sum')


# activity space
exposure_PO(x = geolife_sandiego, cellsize = 50, scale_01 = TRUE,
 input_crs = "EPSG:4326", output_crs = "EPSG:32611", stats = statistics)

#environmental exposure

ndvi_data = terra::rast(system.file("extdata/landsat_ndvi.tif", package = "twsagps"))

exposure_PO(x = geolife_sandiego, cellsize = 50, env_data = ndvi_data,
 scale_01 = TRUE, input_crs = "EPSG:4326", output_crs = "EPSG:32611",
 stats = statistics)




}
