% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KDE_exposure.R
\name{KDE_exposure}
\alias{KDE_exposure}
\title{Kernel Density Estimation exposure (TDA)}
\usage{
KDE_exposure(
  x,
  cellsize = NULL,
  bandwidth = 200,
  env_data = NULL,
  normalize = FALSE,
  data_extent = NULL,
  start_crs = "WGS84",
  end_crs = NULL,
  stats = NULL,
  act_and_env = FALSE
)
}
\arguments{
\item{x}{data frame with lon lat coordinates columns}

\item{cellsize}{positive numeric size of raster cell in meters}

\item{bandwidth}{positive numeric bandwidth in units of x}

\item{env_data}{SpatRaster object of environmental data}

\item{normalize}{boolean argument if activity data should be normalized to 0-1 values range}

\item{data_extent}{TODO}

\item{start_crs}{character or terra crs object specifying coordinate reference system of coordinates in x data frame}

\item{end_crs}{character or terra crs object of coordinate reference system of output}

\item{stats}{vector of characters statistics to be calculated. See terra::global. "count", "range" and "area" additionally added.}
}
\value{
list of SpatRaster result and data frame of statistics
}
\description{
Kernel Density Estimation method activity space and environmental exposure. Using kde() function from TDA package.
Bandwidth is a smoothing parameter same as h in TDA::kde(). In order to receive activity space ignore env_data argument.
}
\examples{

statistics = c("count", "area", "min", "max", "range", "mean", "std", 'sum')

# activity space
KDE_exposure(x = geolife_sandiego, cellsize = 50, bandwidth = 25,
 start_crs = "WGS84", end_crs = "EPSG:32611", stats = statistics)

#environmental exposure
data("landsat_ndvi")
ndvi_data = terra::rast(landsat_ndvi)

KDE_exposure(x = geolife_sandiego, cellsize = 50, bandwidth = 25,
 env_data = ndvi_data, start_crs = "WGS84",
 end_crs = "EPSG:32611", stats = statistics)



}
