% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KDE_exposure.R
\name{KDE_proj}
\alias{KDE_proj}
\title{Kernel Density Estimation in Projected Coordinate Systems}
\usage{
KDE_proj(
  x,
  day = NULL,
  cellsize = 100,
  bandwidth = 200,
  env_data = NULL,
  normalize = FALSE,
  data_extent = NULL,
  start_crs = "WGS84",
  end_crs = NULL,
  stats = NULL,
  act_and_env = FALSE
)
}
\arguments{
\item{x}{data frame with lon lat coordinate columns}

\item{day}{string in date format compatible with date column in x}

\item{cellsize}{size of raster cell in meters}

\item{bandwidth}{bandwidth in meters}

\item{env_data}{SparRaster object of envirinmental data}

\item{normalize}{argument if activity data should be normalized to 0-1 values range}

\item{data_extent}{TODO}

\item{start_crs}{coordinate system of coordinates in x data frame}

\item{end_crs}{coordinate system of output}

\item{stats}{statistics calculated}

\item{act_and_env}{TODO}
}
\value{
list of SpatRaster result and list of statistics
}
\description{
Code working fast with results to check, working well in projected coordinate systems, WGS84 yet to do
}
