% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KDE_wgs.R
\name{KDE_wgs}
\alias{KDE_wgs}
\title{Kernel Density Estimation in WGS84}
\usage{
KDE_wgs(
  x,
  day = NULL,
  cellsize = 100,
  smoothing = 7e-04,
  env_data = NULL,
  data_extent = NULL,
  stats = NULL,
  act_and_env = FALSE
)
}
\arguments{
\item{x}{data frame with lon lat coordinate columns}

\item{day}{string in date format compatible with date column in x}

\item{cellsize}{size of raster cell in meters}

\item{smoothing}{smoothing parameter}

\item{env_data}{SparRaster object of envirinmental data}

\item{data_extent}{TODO}

\item{stats}{statistics calculated}

\item{act_and_env}{TODO}
}
\value{
list of SpatRaster result and list of statistics
}
\description{
Code similar to DR code but using KDE in WGS84, fast but with smoothing parameter
}
