% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{landsat_ndvi}
\alias{landsat_ndvi}
\title{Landsat 9 NDVI image San Diego subset}
\format{
RasterLayer with 390 rows, 304 columns and 1 layer
}
\usage{
data(landsat_ndvi)
}
\description{
San Diego subset of raster NDVI image calculated from Landsat 9 scene
LC90400372024053LGN00 acquired on 22nd February 2024. Raster is in "EPSG:32611" or "WGS84/UTM zone 11N" CRS.
}
\references{
U.S. Geological Survey
}
\keyword{datasets}
