% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LS_exposure.R
\name{LS_exposure}
\alias{LS_exposure}
\title{Line Segment Method exposure}
\usage{
LS_exposure(
  x,
  time_data = NULL,
  time_unit = "mins",
  cellsize = NULL,
  bandwidth = 200,
  env_data = NULL,
  data_extent = NULL,
  start_crs = "WGS84",
  end_crs = NULL,
  stats = NULL,
  act_and_env = FALSE
)
}
\arguments{
\item{x}{data frame with lon lat coordinates columns}

\item{time_data}{name of column in x containing POSIXct data class}

\item{time_unit}{character indicating unit of weights of time_data. Ignored if time_data not specified. See difftime()}

\item{cellsize}{positive numeric size of raster cell in meters}

\item{bandwidth}{positive numeric size of segments in meters}

\item{env_data}{SpatRaster object of environmental data}

\item{data_extent}{TODO}

\item{start_crs}{character or terra crs object specifying coordinate reference system of coordinates in x data frame}

\item{end_crs}{character or terra crs object of coordinate reference system of output}

\item{stats}{vector of characters statistics to be calculated. See terra::global. "count", "range" and "area" additionally added.}
}
\value{
list of SpatRaster result and data frame of statistics
}
\description{
Line Segment method activity space and environmental exposure. Based on code from Micheal garber's repository microclim-static-v-dynam.
In order to receive activity space ignore env_data argument.
}
\examples{

statistics = c("count", "area", "min", "max", "range", "mean", "std", 'sum')

# activity space
LS_exposure(x = geolife_sandiego, time_data = dateTime, time_unit = "mins",
 cellsize = 50, bandwidth = 200, start_crs = "WGS84", end_crs = "EPSG:32611",
  stats = statistics)

#environmental exposure
data("landsat_ndvi")
ndvi_data = terra::rast(landsat_ndvi)

LS_exposure(x = geolife_sandiego, time_data = dateTime, time_unit = "mins",
 cellsize = 50, bandwidth = 200, env_data = ndvi_data, start_crs = "WGS84",
 end_crs = "EPSG:32611", stats = statistics)


}
