% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PO_exposure.R
\name{PO_exposure}
\alias{PO_exposure}
\title{Point Overlay exposure}
\usage{
PO_exposure(
  x,
  day = NULL,
  cellsize = 100,
  env_data = NULL,
  normalize = FALSE,
  data_extent = NULL,
  start_crs = "WGS84",
  end_crs = NULL,
  stats = NULL,
  act_and_env = FALSE
)
}
\arguments{
\item{x}{data frame with lon lat coordinate columns}

\item{day}{string in date format compatible with date column in x}

\item{cellsize}{size of raster cell in meters}

\item{env_data}{SpatRaster object of envirinmental data}

\item{normalize}{argument if activity data should be normalized to 0-1 values range}

\item{data_extent}{TODO}

\item{start_crs}{coordinate system of coordinates in x data frame}

\item{end_crs}{coordinate system of output}

\item{stats}{statistics calculated}

\item{act_and_env}{TODO}
}
\value{
list of SpatRaster result and list of statistics
}
\description{
Point Overlay exposure
}
