% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rast_stats.R
\name{rast_stats}
\alias{rast_stats}
\title{Raster statistics}
\usage{
rast_stats(..., stats, row_names = NULL)
}
\arguments{
\item{...}{SpatRaster objects to calculate statistics on}

\item{stats}{vector of statistics to be calculated. See terra::global() for more details. Added "range", "count" and "area".}

\item{row_names}{vector the same length as number of ... arguments specifying row names of output data.frame. If NULL row names are ... arguments, values without argument specified will be numbered row-wise.}
}
\value{
Data.frame of length(stats) columns and ... rows
}
\description{
Calculate statistics dataframe from multiple SpatRasters
}
\examples{
statistics = c("count", "area", "min", "max", "range", "mean", "std", 'sum')

ndvi_data = terra::rast(system.file("extdata/landsat_ndvi.tif", package = "twiGPS"))

exposure = exposure_PO(data = geolife_sandiego, x = lon, y = lat, cellsize = 50, normalize = TRUE,
                       input_crs = "EPSG:4326", output_crs = "EPSG:32611")
# no names
rast_stats(ndvi_data, exposure, stats = statistics)

# argument names
rast_stats(ndvi = ndvi_data, PO = exposure, stats = statistics)

# names
rast_stats(ndvi_data, exposure, stats = statistics, row_names = c("NDVI", "PO"))

}
