% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twiGPS-package.R
\docType{package}
\name{twiGPS-package}
\alias{twiGPS}
\alias{twiGPS-package}
\title{twiGPS: Time-Weighted Spatial Averaging Methods with GPS}
\description{
Time-Weighted Spatial Averaging (TWSA) methods for calculating activity space and environmental exposure from GPS data. This package implements four TWSA methods - Point Overlay, Kernel Density Estimation, Density Ranking (Jankowska et al., 2023 <doi.org/10.1016/j.healthplace.2021.102706>) with the addtion of Line Segment method (github.com/michaeldgarber/microclim-static-v-dynam).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Graczykowski/twiGPS}
  \item Report bugs at \url{https://github.com/Graczykowski/twiGPS/issues}
}

}
\author{
\strong{Maintainer}: Igor Graczykowski \email{igor.graczykowski@gmail.com}

Authors:
\itemize{
  \item Jakub Nowosad \email{nowosad.jakub@gmail.com} (\href{https://orcid.org/0000-0002-1057-3721}{ORCID})
  \item Marta Jankowska \email{mjankowska@coh.org} (\href{https://orcid.org/0000-0002-8688-1181}{ORCID})
  \item Jay Yang \email{jyang@coh.org} (\href{https://orcid.org/0000-0003-4246-0470}{ORCID})
}

}
\keyword{internal}
