% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure_DR.R
\name{exposure_DR}
\alias{exposure_DR}
\title{Density Ranking exposure}
\usage{
exposure_DR(
  data,
  x,
  y,
  NA_val,
  cellsize,
  group_split,
  bandwidth = 200,
  env_data,
  env_field,
  env_buff,
  normalize = FALSE,
  norm_method = "range",
  norm_group = FALSE,
  grid_extent,
  input_crs,
  output_crs,
  filepath
)
}
\arguments{
\item{data}{Data.frame, SpatVector points or sf data.frame containing only POINTS.}

\item{x}{Data masking. x or lontitude coordinates column if data is a data.frame.}

\item{y}{Data masking. y or latitude coordinates column if data is a data.frame.}

\item{NA_val}{Numeric. Value in x and y marked as NA if data is a data.frame.}

\item{cellsize}{Positive numeric. Size of raster cells in meters if output has a longtitude/latitude CRS or in the units of output CRS.}

\item{group_split}{Data masking. Column of data based on which it is grouped and split. For n groups output will be n SpatRasters.}

\item{bandwidth}{Positive numeric. Bandwidth in units of output Coordinate Reference System.}

\item{env_data}{Stars, SpatRaster, SpatVector or sf. Spatial environmental data. Activity space is calculated when not set. When argument is a SpatVector or sf object vector data is rasterized to output raster using "sum" function.}

\item{env_field}{Data masking. Column of env_data that env_data will be rasterized on. Ignored if env_data not a SpatVector or sf class.}

\item{env_buff}{Positive numeric. Optional buffer around SpatVector/sf env_data in meters if output has a longtitude/latitude CRS or in the units of the CRS. See terra::buffer(). Ignored if env_data not a SpatVector or sf class.}

\item{normalize}{Boolean. If activity data should be normalized.}

\item{norm_method}{Character. Normalization method. Four methods - "center", "scale", "standardize" and "range". Default is "range". See BBmisc::normalize().}

\item{norm_group}{Boolean. When normalize is TRUE, norm_method is "range" and group_split is set. If FALSE each SpatRaster is rescaled seperately. If TRUE SpatRasters are rescaled to SpatRaster with highest max value.}

\item{grid_extent}{Stars, SpatRaster, SpatExtent, sf bbox object or numeric vector of 4 length c(xmin, xmax, ymin, ymax). If stars or SpatRaster grid_extent is output's grid and cellsize argument is ignored. If SpatExtent, sf bbox object or vector grid_extent is output's extent. If cellsize is set it is preserved at the cost of extent.}

\item{input_crs}{Character or terra crs object. Coordinate Reference System of data's coordinates if data is a data.frame.}

\item{output_crs}{Character or terra crs object. Coordinate Reference System of output. If not set and env_data is a SpatRaster env_data's CRS is used.}

\item{filepath}{Character. Output filename. See terra::writeRaster().}
}
\value{
SpatRaster
}
\description{
Density Ranking method activity space and environmental exposure.
Using spat_dr() function based on Silverman, B. W. 1986 (KDE) and DR function from Yen-Chi Chen's repository yenchic/density_ranking.
In order to receive activity space ignore env_data argument.
}
\examples{

# activity space
exposure_DR(data = geolife_sandiego, x = lon, y = lat, cellsize = 50, bandwidth = 200,
 normalize = TRUE, input_crs = "EPSG:4326", output_crs = "EPSG:32611")

# split by date
exposure_DR(data = geolife_sandiego, x = lon, y = lat, cellsize = 50, group_split = date,
   bandwidth = 200, normalize = TRUE, norm_group = FALSE, input_crs = "EPSG:4326",
   output_crs = "EPSG:32611")

# split by date and define extent
extent = c(478000, 484000, 3618000, 3627000)

exposure_DR(data = geolife_sandiego, x = lon, y = lat, cellsize = 50, group_split = date,
  bandwidth = 200, normalize = TRUE,norm_group = TRUE, grid_extent = extent,
  input_crs = "EPSG:4326", output_crs = "EPSG:32611")

#environmental exposure

ndvi_data = terra::rast(system.file("extdata/landsat_ndvi.tif", package = "twsagps"))

exposure_DR(data = geolife_sandiego, x = lon, y = lat, cellsize = 50, bandwidth = 200,
  env_data = ndvi_data, normalize = FALSE, input_crs = "EPSG:4326", output_crs = "EPSG:32611")

# environmental exposure - use rast grid and split by date
exposure_DR(data = geolife_sandiego, x = lon, y = lat, group_split = date,
  bandwidth = 200, env_data = ndvi_data, normalize = TRUE, norm_group = TRUE,
  grid_extent = ndvi_data, input_crs = "EPSG:4326", output_crs = "EPSG:32611")

}
