% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure_LS.R
\name{exposure_LS}
\alias{exposure_LS}
\title{Line Segment Method exposure}
\usage{
exposure_LS(
  data,
  coords,
  bandwidth,
  cellsize,
  time_data,
  env_data,
  output_crs,
  input_crs,
  grid_extent,
  normalize,
  group_split,
  norm_group = FALSE,
  filepath,
  NA_val,
  time_unit = "mins",
  env_field,
  env_buff,
  verbose = TRUE
)
}
\arguments{
\item{data}{Data.frame, SpatVector points or sf data.frame containing only POINTS.}

\item{coords}{Character. Vector of column names of x and y coordinates if data is a data.frame.}

\item{bandwidth}{Positive numeric. Bandwidth in meters if output has a longtitude/latitude CRS or in the units of output CRS.}

\item{cellsize}{Positive numeric. Size of raster cells in meters if output has a longtitude/latitude CRS or in the units of output CRS.}

\item{time_data}{Data masking. Column of data containing POSIXct class objects.}

\item{env_data}{Stars, SpatRaster, SpatVector or sf. Spatial environmental data. Activity space is calculated when not set. When argument is a SpatVector or sf object vector data is rasterized to output raster using "sum" function.}

\item{output_crs}{Character or terra crs object. Coordinate Reference System of output. If not set and env_data is a SpatRaster env_data's CRS is used.}

\item{input_crs}{Character or terra crs object. Coordinate Reference System of data's coordinates if data is a data.frame.}

\item{grid_extent}{Stars, SpatRaster, SpatExtent, sf bbox object or numeric vector of 4 length c(xmin, xmax, ymin, ymax). If stars or SpatRaster grid_extent is output's grid and cellsize argument is ignored. If SpatExtent, sf bbox object or vector grid_extent is output's extent. If cellsize is set it is preserved at the cost of extent.}

\item{normalize}{Character. If not empty activity space SpatRaster is normalized with specified method. Four methods - "center", "scale", "standardize" and "range". Default is "range". See BBmisc::normalize().}

\item{group_split}{Data masking. Column of data based on which it is grouped and split. For n groups output will be n SpatRasters.}

\item{norm_group}{Boolean. When normalize is TRUE, norm_method is "range" and group_split is set. If FALSE each SpatRaster is normalized seperately. If TRUE SpatRasters are normalized to SpatRaster with highest max value.}

\item{filepath}{Character. Output filename. See terra::writeRaster().}

\item{NA_val}{Numeric. Value in x and y marked as NA if data is a data.frame.}

\item{time_unit}{Character. Unit of time weights of time_data. Ignored if time_data not specified. See difftime().}

\item{env_field}{Data masking. Column of env_data that env_data will be rasterized on. Ignored if env_data not a SpatVector or sf class.}

\item{env_buff}{Positive numeric. Optional buffer around SpatVector/sf env_data in meters if output has a longtitude/latitude CRS or in the units of the CRS. See terra::buffer(). Ignored if env_data not a SpatVector or sf class.}

\item{verbose}{Boolean. If TRUE amount of output is reduced.}
}
\value{
SpatRaster
}
\description{
Line Segment method activity space and environmental exposure. Based on code from Micheal Garber's repository michaeldgarber/microclim-static-v-dynam.
In order to receive activity space ignore env_data argument.

For other exposure methods see: \code{\link[=exposure_PO]{exposure_PO()}}, \code{\link[=exposure_KDE]{exposure_KDE()}}, \code{\link[=exposure_DR]{exposure_DR()}}

For calculating statistics of output raster use \code{\link[=exposure_stats]{exposure_stats()}}
}
\examples{
# activity space
exposure_LS(data = geolife_sandiego, coords = c("lon", "lat"), bandwidth = 200,
  cellsize = 50, time_data = dateTime, output_crs = "EPSG:32611",
  input_crs = "EPSG:4326", normalize = "range", verbose = FALSE)

# split by date
exposure_LS(data = geolife_sandiego, coords = c("lon", "lat"), bandwidth = 200,
  cellsize = 50, normalize = "range", group_split = date, norm_group = FALSE,
  output_crs = "EPSG:32611", input_crs = "EPSG:4326", verbose = FALSE)

# split by date and define extent with time_only_norm
extent = c(478000, 484000, 3618000, 3627000)

exposure_LS(data = geolife_sandiego, coords = c("lon", "lat"), bandwidth = 200,
  cellsize = 50, time_data = dateTime, output_crs = "EPSG:32611",
  input_crs = "EPSG:4326", grid_extent = extent, normalize = "range",
  group_split = date, norm_group = TRUE, verbose = FALSE)

#environmental exposure

ndvi_data = terra::rast(system.file("extdata/landsat_ndvi.tif", package = "twiGPS"))

exposure_LS(data = geolife_sandiego, coords = c("lon", "lat"), bandwidth = 200,
  cellsize = 50, time_data = dateTime, env_data = ndvi_data,
  output_crs = "EPSG:32611", input_crs = "EPSG:4326", verbose = FALSE)

# environmental exposure - use rast grid and split by date with time_only_norm
exposure_LS(data = geolife_sandiego, coords = c("lon", "lat"), bandwidth = 200,
  cellsize = 50, time_data = dateTime, env_data = ndvi_data,
  output_crs = "EPSG:32611", input_crs = "EPSG:4326", grid_extent = ndvi_data,
  normalize = "range", group_split = date, norm_group = TRUE, verbose = FALSE)
}
