% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure_LS.R
\name{exposure_LS}
\alias{exposure_LS}
\title{Line Segment Method exposure}
\usage{
exposure_LS(
  data,
  x,
  y,
  NA_val,
  time_data,
  time_unit = "mins",
  cellsize,
  group_split,
  bandwidth,
  env_data,
  env_field,
  env_buff,
  normalize = FALSE,
  norm_method = "range",
  norm_group = FALSE,
  time_only_norm = FALSE,
  grid_extent,
  input_crs,
  output_crs,
  filepath
)
}
\arguments{
\item{data}{Data.frame, SpatVector points or sf data.frame containing only POINTS.}

\item{x}{Data masking. x or longitude coordinates column name if data is a data.frame.}

\item{y}{Data masking. y or latitude coordinates column name if data is a data.frame.}

\item{NA_val}{Numeric. Value in x and y marked as NA if data is a data.frame.}

\item{time_data}{Data masking. Column of data containing POSIXct class objects.}

\item{time_unit}{Character. Unit of time weights of time_data. Ignored if time_data not specified. See difftime().}

\item{cellsize}{Positive numeric. Size of raster cells in meters if output has a longtitude/latitude CRS or in the units of output CRS.}

\item{group_split}{Data masking. Column of data based on which it is grouped and split. For n groups output will be n SpatRasters.}

\item{bandwidth}{Positive numeric. Bandwidth in meters if output has a longtitude/latitude CRS or in the units of output CRS.}

\item{env_data}{Stars, SpatRaster, SpatVector or sf. Spatial environmental data. Activity space is calculated when not set. When argument is a SpatVector or sf object vector data is rasterized to output raster using "sum" function.}

\item{env_field}{Data masking. Column of env_data that env_data will be rasterized on. Ignored if env_data not a SpatVector or sf class.}

\item{env_buff}{Positive numeric. Optional buffer around SpatVector/sf env_data in meters if output has a longtitude/latitude CRS or in the units of the CRS. See terra::buffer(). Ignored if env_data not a SpatVector or sf class.}

\item{normalize}{Boolean. If TRUE activity space SpatRaster is normalized}

\item{norm_method}{Character. Normalization method. Four methods - "center", "scale", "standardize" and "range". Default is "range". See BBmisc::normalize().}

\item{norm_group}{Boolean. When normalize is TRUE, norm_method is "range" and group_split is set. If FALSE each SpatRaster is normalized seperately. If TRUE SpatRasters are normalized to SpatRaster with highest max value.}

\item{time_only_norm}{Boolean. Time type of normalization. If TRUE only time_data is normalized, normalized raster values may be imprecise. If FALSE and norm_method is "range" normalization is conducted with regard to spatial distribution, normalized raster values are accurate. TEMPORARY}

\item{grid_extent}{Stars, SpatRaster, SpatExtent, sf bbox object or numeric vector of 4 length c(xmin, xmax, ymin, ymax). If stars or SpatRaster grid_extent is output's grid and cellsize argument is ignored. If SpatExtent, sf bbox object or vector grid_extent is output's extent. If cellsize is set it is preserved at the cost of extent.}

\item{input_crs}{Character or terra crs object. Coordinate Reference System of data's coordinates if data is a data.frame.}

\item{output_crs}{Character or terra crs object. Coordinate Reference System of output. If not set and env_data is a SpatRaster env_data's CRS is used.}

\item{filepath}{Character. Output filename. See terra::writeRaster().}
}
\value{
SpatRaster
}
\description{
Line Segment method activity space and environmental exposure. Based on code from Micheal Garber's repository michaeldgarber/microclim-static-v-dynam.
In order to receive activity space ignore env_data argument.
}
\examples{

# activity space
exposure_LS(data = geolife_sandiego, x = lon, y = lat, time_data = dateTime, time_unit = "mins",
  cellsize = 50, bandwidth = 200, normalize = TRUE, input_crs = "EPSG:4326", output_crs = "EPSG:32611")

# split by date
exposure_LS(data = geolife_sandiego, x = lon, y = lat, cellsize = 50, group_split = date,
   bandwidth = 200, normalize = TRUE, norm_group = FALSE, input_crs = "EPSG:4326",
   output_crs = "EPSG:32611")

# split by date and define extent with time_only_norm
extent = c(478000, 484000, 3618000, 3627000)

exposure_LS(data = geolife_sandiego, x = lon, y = lat, time_data = dateTime, time_unit = "mins",
  cellsize = 50, group_split = date, bandwidth = 200, normalize = TRUE, norm_group = TRUE,
  time_only_norm = TRUE, grid_extent = extent, input_crs = "EPSG:4326",
  output_crs = "EPSG:32611")

#environmental exposure

ndvi_data = terra::rast(system.file("extdata/landsat_ndvi.tif", package = "twsagps"))

exposure_LS(data = geolife_sandiego, x = lon, y = lat, time_data = dateTime, time_unit = "mins",
  cellsize = 50, bandwidth = 200, env_data = ndvi_data, normalize = FALSE,
  input_crs = "EPSG:4326", output_crs = "EPSG:32611")

# environmental exposure - use rast grid and split by date with time_only_norm
exposure_LS(data = geolife_sandiego, x = lon, y = lat, time_data = dateTime, time_unit = "mins",
  group_split = date, bandwidth = 200, env_data = ndvi_data, normalize = TRUE,
  norm_group = TRUE, time_only_norm = TRUE, grid_extent = ndvi_data, input_crs = "EPSG:4326",
  output_crs = "EPSG:32611")
}
