% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure_LS.R
\name{exposure_LS}
\alias{exposure_LS}
\title{Calculate Line Segment exposure of point spatial data}
\usage{
exposure_LS(
  data,
  coords,
  bandwidth,
  cellsize,
  time_data,
  env_data,
  output_crs,
  input_crs,
  grid_extent,
  normalize,
  group_split,
  norm_group = FALSE,
  filepath,
  NA_val,
  time_unit = "mins",
  env_field,
  env_buff,
  verbose = TRUE
)
}
\arguments{
\item{data}{Data.frame, SpatVector points or sf data.frame containing only POINTS.}

\item{coords}{Character. Vector of column names of x and y coordinates if \code{data} is a data.frame.}

\item{bandwidth}{Positive numeric. Bandwidth in meters if output has a longtitude/latitude CRS or in the units of output's CRS.}

\item{cellsize}{Positive numeric. Size of raster cells in meters if output has a longtitude/latitude CRS or in the units of output CRS.}

\item{time_data}{Data masking. Column of data containing POSIXct class objects.}

\item{env_data}{Stars, SpatRaster, SpatVector or sf. Spatial environmental data. Activity space is calculated when not set. When argument is a SpatVector or sf object vector data is rasterized to output raster using \code{"sum"} function.}

\item{output_crs}{Character or terra crs object. CRS of output. If not set and \code{env_data} is a SpatRaster \code{env_data}'s CRS is used.}

\item{input_crs}{Character or terra crs object. CRS of \code{data}'s coordinates if \code{data} is a data.frame.}

\item{grid_extent}{Stars, SpatRaster, SpatExtent, sf bbox object or numeric vector of 4 length \code{c(xmin, xmax, ymin, ymax)}. If stars or SpatRaster \code{grid_extent} is output's grid and \code{cellsize} argument is ignored. If SpatExtent, sf bbox object or vector \code{grid_extent} is output's extent. Then, if \code{cellsize} is set it is preserved at the cost of extent.}

\item{normalize}{Character. If set activity space SpatRaster is normalized with specified method. Four methods - "center", "scale", "standardize" and "range". See \code{\link[BBmisc:normalize]{BBmisc::normalize()}}.}

\item{group_split}{Data masking. Column of data based on which it is grouped and split. For n groups output will be SpatRasters with n layers.}

\item{norm_group}{Boolean. Applicable only When normalize is "range" and group_split is set. If FALSE each layer of SpatRaster is normalized seperately. If TRUE all layer are normalized to layer with highest max value (default FALSE).}

\item{filepath}{Character. Output filename. See \code{\link[terra:writeRaster]{terra::writeRaster()}}.}

\item{NA_val}{Numeric. Value in x and y marked as NA if \code{data} is a data.frame.}

\item{time_unit}{Character. Unit of time weights of \code{time_data} (default \code{"mins"}). Ignored if \code{time_data} not specified. See \code{\link[=difftime]{difftime()}}.}

\item{env_field}{Data masking. Column of \code{env_data} that \code{env_data} will be rasterized on. Ignored if \code{env_data} not a SpatVector or sf class.}

\item{env_buff}{Positive numeric. Optional buffer around SpatVector/sf \code{env_data} in meters if output has a longtitude/latitude CRS or in the units of the CRS. Ignored if \code{env_data} not a SpatVector or sf class. See \code{\link[terra:buffer]{terra::buffer()}}.}

\item{verbose}{Boolean. If FALSE amount of output is reduced (default TRUE).}
}
\value{
SpatRaster
}
\description{
\code{exposure_LS()} calculates activity space and environmental exposure using Line Segment method.
Line Segment function is based on code from Micheal Garber's repository michaeldgarber/microclim-static-v-dynam.

This method creates trajectories following points of spatial data and creates buffers around each trajectory.
Subsequently, values for every buffer are extracted from \code{env_data}.
Finally, \code{time_data} weights are incorporated and converted to SpatRaster with \code{\link[terra:rasterize]{terra::rasterize()}} by \code{fun = "sum"}.
Setting \code{env_data} argument determines whether environmental exposure or activity space is computed.

For calculating statistics of output raster use \code{\link[=exposure_stats]{exposure_stats()}}
}
\examples{
# activity space for data.frame 'data' wihout time data
exposure_LS(data = geolife_sandiego, coords = c("lon", "lat"),
  bandwidth = 200, cellsize = 50, output_crs = "EPSG:32611",
  input_crs = "EPSG:4326")

#SpatVector data
geolife_vect = terra::vect(geolife_sandiego,
                           geom = c("lon", "lat"),
                           crs = "EPSG:4326") |>
                           terra::project("EPSG:32611")

# time_data
exposure_LS(data = geolife_vect, bandwidth = 200,
  cellsize = 50, time_data = dateTime)

# normalize "range"
exposure_LS(data = geolife_vect, bandwidth = 200,
  cellsize = 50, time_data = dateTime, normalize = "range")

# normalize "center"
exposure_LS(data = geolife_vect, bandwidth = 200,
  cellsize = 50, time_data = dateTime, normalize = "center")

# split by date
exposure_LS(data = geolife_vect, bandwidth = 200,
  cellsize = 50, time_data = dateTime,
  normalize = "range", group_split = date)

# split by date, normalize by group and define extent
extent = c(478000, 484000, 3618000, 3627000)

exposure_LS(data = geolife_vect, bandwidth = 200,
  cellsize = 50, time_data = dateTime,
  grid_extent = extent, normalize = "range",
  group_split = date, norm_group = TRUE)

# environmental exposure
ndvi_data = terra::rast(system.file("extdata/landsat_ndvi.tif",
                                     package = "twiGPS"))

exposure_LS(data = geolife_vect, bandwidth = 200,
 cellsize = 50, time_data = dateTime, env_data = ndvi_data)

# environmental exposure - use rast grid and split by date
exposure_LS(data = geolife_vect, bandwidth = 200,
  time_data = dateTime, env_data = ndvi_data,
  grid_extent = ndvi_data, normalize = "range",
  group_split = date)
}
\references{
Jankowska, Marta M., Jiue-An Yang, Nana Luo, Chad Spoon, and Tarik Benmarhnia. 2023. “Accounting for Space, Time, and Behavior Using GPS Derived Dynamic Measures of Environmental Exposure.” Health & Place 79 (January): 102706. https://doi.org/10.1016/j.healthplace.2021.102706.
}
\seealso{
\code{\link[=exposure_PO]{exposure_PO()}}, \code{\link[=exposure_KDE]{exposure_KDE()}}, \code{\link[=exposure_DR]{exposure_DR()}}
}
