% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure_stats.R
\name{exposure_stats}
\alias{exposure_stats}
\title{Exposure statistics}
\usage{
exposure_stats(..., stats, row_names = NULL, verbose = TRUE)
}
\arguments{
\item{...}{SpatRaster objects from which statistics are calculated}

\item{stats}{Character. Single or vector of statistics to be calculated. See \code{\link[terra:global]{terra::global()}} for more details. Added "range", "count" and "area".}

\item{row_names}{Character. Single or vector the same length as number of ... arguments specifying row names of output data.frame. If NULL row names are ... arguments, values without argument specified will be numbered row-wise. If ... SpatRasters have more than one layer row_names vector should be a length of sum of layers of all SpatRasters.}

\item{verbose}{Boolean. If TRUE amount of output is reduced.}
}
\value{
Data.frame of \code{length(stats)} columns and \code{length(...)} rows
}
\description{
Calculate statistics data.frame from multiple SpatRasters
}
\examples{
statistics = c("count", "area", "min", "max", "range", "mean", "std", 'sum')

ndvi_data = terra::rast(system.file("extdata/landsat_ndvi.tif",
                                     package = "twiGPS"))

exposure = exposure_PO(data = geolife_sandiego, coords = c("lon", "lat"),
                       cellsize = 50, input_crs = "EPSG:4326", output_crs = "EPSG:32611")
# no names
exposure_stats(ndvi_data, exposure, stats = statistics)

# argument names
exposure_stats(ndvi = ndvi_data, PO = exposure, stats = statistics)

# names
exposure_stats(ndvi_data, exposure, stats = statistics,
 row_names = c("NDVI", "Point Exposure"))
}
\seealso{
\code{\link[=exposure_PO]{exposure_PO()}}, \code{\link[=exposure_KDE]{exposure_KDE()}}, \code{\link[=exposure_DR]{exposure_DR()}}, \code{\link[=exposure_LS]{exposure_LS()}}
}
