% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LS_exposure.R
\name{LS_exposure}
\alias{LS_exposure}
\title{Line Segment Method exposure}
\usage{
LS_exposure(
  x,
  day = NULL,
  time_data = NULL,
  cellsize = 100,
  buff_dist = 200,
  env_data = NULL,
  data_extent = NULL,
  start_crs = "WGS84",
  end_crs = NULL,
  stats = NULL,
  act_and_env = FALSE
)
}
\arguments{
\item{x}{data frame with lon lat coordinate columns}

\item{day}{string in date format compatible with date column in x}

\item{time_data}{name of column in x containing POSIXct data class}

\item{cellsize}{size of raster cell in meters}

\item{buff_dist}{size of buffer im meter}

\item{env_data}{SparRaster object of envirinmental data}

\item{data_extent}{TODO}

\item{start_crs}{coordinate system of coordinates in x data frame}

\item{end_crs}{coordinate system of output}

\item{stats}{statistics calculated}

\item{act_and_env}{TODO}
}
\value{
list of SpatRaster result and list of statistics
}
\description{
Based on code from Micheal garber's repository microclim-static-v-dynam
}
