% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PO_exposure.R
\name{PO_exposure}
\alias{PO_exposure}
\title{Point Overlay exposure}
\usage{
PO_exposure(
  x,
  cellsize = NULL,
  env_data = NULL,
  normalize = FALSE,
  data_extent = NULL,
  start_crs = "WGS84",
  end_crs = NULL,
  stats = NULL,
  act_and_env = FALSE
)
}
\arguments{
\item{x}{data frame with lon lat coordinates columns}

\item{cellsize}{positive numeric size of raster cell in meters}

\item{env_data}{SpatRaster object of environmental data}

\item{normalize}{boolean argument if activity data should be normalized to 0-1 values range}

\item{data_extent}{TODO}

\item{start_crs}{character or terra crs object specifying coordinate reference system of coordinates in x data frame}

\item{end_crs}{character or terra crs object of coordinate reference system of output}

\item{stats}{vector of characters statistics to be calculated. See terra::global. "count", "range" and "area" additionally added.}
}
\value{
list of SpatRaster result and data frame of statistics
}
\description{
Point Overlay method activity space and environmental exposure. In order to receive activity space ignore env_data argument.
}
\examples{

statistics = c("count", "area", "min", "max", "range", "mean", "std", 'sum')


# activity space
PO_exposure(x = geolife_sandiego, cellsize = 50, normalize = TRUE,
 start_crs = "WGS84", end_crs = "EPSG:32611", stats = statistics)

#environmental exposure

ndvi_data = terra::rast(system.file("extdata/landsat_ndvi.tif", package = "twsagps"))

PO_exposure(x = geolife_sandiego, cellsize = 50, env_data = ndvi_data,
 normalize = TRUE, start_crs = "WGS84", end_crs = "EPSG:32611",
 stats = statistics)




}
