% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twiGPS-package.R
\docType{package}
\name{twiGPS-package}
\alias{twiGPS}
\alias{twiGPS-package}
\title{twiGPS: Time-Weighted spatial averaging methods Integrated with GPS measured environmental exposure}
\description{
TWSA methods (Point Overlay, Kernel Density Estimation, Density Ranking, Line Segment) Integrated with GPS measured environmental exposure. Sample data - geolife_sandiego (sample data from Geolife GPS Trajectories 1.3 Data for San Diego county). Environmental data - landsat_ndvi (NDVI calculated from Landsat 9 image for San Diego county captured on 22nd Februrary 2024).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Graczykowski/twiGPS}
  \item Report bugs at \url{https://github.com/Graczykowski/twiGPS/issues}
}

}
\author{
\strong{Maintainer}: Igor Graczykowski \email{igor.graczykowski@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
